Discord Emoji Downloader/requirements.txt
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.34.2
Nuitka==2.4.5
numpy==2.0.1
ordered-set==4.1.0
pillow==10.4.0
pip==24.2
pycryptodome==3.20.0
pypiwin32==223
pyreadline3==3.4.1
pywin32==306
requests==2.32.3
setuptools==72.1.0
sniffio==1.3.1
urllib3==2.2.2
zstandard==0.23.0
